+++ test/not_loaded_test.exs
@@ -10,6 +10,15 @@ defmodule NotLoadedTest do
     ]
   end
 
+  defmodule Employee do
+    import NotLoadedTest.CustomNotLoaded, only: [lazy_loaded: 1]
+
+    defstruct [
+      :name,
+      lazy_loaded(:address)
+    ]
+  end
+
   test "inspect" do
     book = %Book{}
 
@@ -18,5 +27,17 @@ defmodule NotLoadedTest do
 
   test "new/1 creates a new not_loaded struct" do
     assert %NotLoaded{__field__: :password} = NotLoaded.new(:password)
+  end
+
+  test "with a custom NotLoaded module" do
+    employee = %Employee{}
+
+    not_loaded_struct = employee.address
+
+    assert "#NotLoadedTest.CustomNotLoaded<field :address is not loaded>" =
+             inspect(not_loaded_struct)
+
+    assert not_loaded_struct.__owner__ == Employee
+    assert not_loaded_struct.__field__ == :address
   end
 end