+++ test/support/custom_modules.ex
@@ -1,0 +1,1 @@
+NotLoaded.defmodule(NotLoadedTest.CustomNotLoaded)